Code review feedback, playing with Transaction.
